Make use of the rich package; insert button to compute median

+++ src/teareduce/cookbook/get_cookbook_file.py
@@ -7,6 +7,7 @@
 # License-Filename: LICENSE.txt
 
 import os
+from pathlib import Path
 import requests
 
 
@@ -28,6 +29,10 @@ def get_cookbook_file(file_path, verbose=True):
     url = f"https://raw.githubusercontent.com/nicocardiel/teareduce-cookbook/main/{file_path}"
     response = requests.get(url)
     local_filename = os.path.basename(file_path)
+    if Path(local_filename).is_file():
+        if verbose:
+            print(f"File {local_filename} already exists locally. Skipping download.")
+        return
     with open(local_filename, "wb") as f:
         f.write(response.content)
     if verbose:
